(feat) Replace Assoc with Key and use associated types.
This massively improves the ergonomics of TypeMap operations
by only requiring the key type to be specified for lookups
rather than both types.

This is, of course, a [breaking-change]. To migrate, simply
change all implementations of Assoc to implementations of Key,
where the type parameter is replaced with an associated type
of Value.

Additionally, all invocations of typemap methods now require
only a single type parameter and all entry types now have
only one type parameter for the key, rather than two for
both the key and the value.