(feat) Add CloneMap, a variant of TypeMap which is Clone.
Getting the right things to be Clone is tricky, but that's just
because Clone isn't object safe and we don't have multi-trait
trait objects.

Also, we have to be a bit cleverer than derive here by bounding
by `Box<A>: Clone`, *not* `A: Clone`, so that we can support our
non-sized `CloneAny` type. Note that `Box<CloneAny>` implements
`Clone`, but `CloneAny` doesn't.