(feat) Add Implements to truly allow custom TypeMaps.
The previous addition of A was not enough, since the implementation
block only applied to A = UnsafeAny.

The implementation now works by a manual trait, Implements, which is
used in bounds. This has the unfortunate consequence of preventing
new TypeMap kinds from being defined outside of this crate. Hopefully,
future type system extensions will allow this restriction to be lifted.